# Improve performance of rebuilds