@RobinMalfait 2025-02-05
@philipp-spiess haha just a nice feeling when you move some code around and improve performance 😅